Move httpCommandServer.ts into main/commandServer
