+++ src/main/commandServer/httpCommandServer.ts
